Er_mostro/llama3_rag_setup/etl/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
python-dotenv 
loguru
python-multipart